Modificación ciudad, tipo documento, fecha nacimiento.
Se implementa también lombok.

+++ src/main/java/com/talentpool/monolito/repository/IClienteRepository.java
@@ -3,5 +3,5 @@ package com.talentpool.monolito.repository;
 import com.talentpool.monolito.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface IClienteRepository extends JpaRepository<Cliente, Long> {
 }
